+++ lib/src/app.dart
@@ -414,9 +414,11 @@ class _FluentAppState extends State<FluentApp> {
   Widget _builder(BuildContext context, Widget? child) {
     final themeData = theme(context);
     final mTheme = context.findAncestorWidgetOfExactType<m.Theme>();
+
     return m.AnimatedTheme(
       data: mTheme?.data ??
           m.ThemeData(
+            extensions: themeData.extensions.values,
             brightness: themeData.brightness,
             canvasColor: themeData.cardColor,
             textSelectionTheme: TextSelectionThemeData(
@@ -425,22 +427,6 @@ class _FluentAppState extends State<FluentApp> {
                   .withOpacity(0.8),
               cursorColor: themeData.inactiveColor,
             ),
-            // colorScheme: m.ColorScheme.fromSwatch(
-            //   primarySwatch: m.MaterialColor(
-            //     400,
-            //     {
-            //       100: themeData.accentColor.lightest,
-            //       200: themeData.accentColor.lighter,
-            //       300: themeData.accentColor.light,
-            //       400: themeData.accentColor.normal,
-            //       500: themeData.accentColor.dark,
-            //       600: themeData.accentColor.darker,
-            //       700: themeData.accentColor.darkest,
-            //     },
-            //   ),
-            //   primaryColorDark: themeData.accentColor.dark,
-            //   brightness: themeData.brightness,
-            // ),
           ),
       child: AnimatedFluentTheme(
         curve: themeData.animationCurve,