+++ lib/src/styles/theme.dart
@@ -175,6 +175,7 @@ const standardCurve = Curves.easeInOut;
 @immutable
 class ThemeData with Diagnosticable {
   final Typography typography;
+  final Map<Object, ThemeExtension<dynamic>> extensions;
 
   final AccentColor accentColor;
   final Color activeColor;
@@ -225,6 +226,7 @@ class ThemeData with Diagnosticable {
   final ResourceDictionary resources;
 
   factory ThemeData({
+    Iterable<ThemeExtension<dynamic>>? extensions,
     Brightness? brightness,
     VisualDensity? visualDensity,
     Typography? typography,
@@ -271,6 +273,7 @@ class ThemeData with Diagnosticable {
     ResourceDictionary? resources,
   }) {
     brightness ??= Brightness.light;
+    extensions ??= [];
 
     final isLight = brightness == Brightness.light;
 
@@ -343,6 +346,7 @@ class ThemeData with Diagnosticable {
 
     return ThemeData.raw(
       brightness: brightness,
+      extensions: _themeExtensionIterableToMap(extensions),
       visualDensity: visualDensity,
       fasterAnimationDuration: fasterAnimationDuration,
       fastAnimationDuration: fastAnimationDuration,
@@ -390,6 +394,7 @@ class ThemeData with Diagnosticable {
 
   const ThemeData.raw({
     required this.typography,
+    required this.extensions,
     required this.accentColor,
     required this.activeColor,
     required this.inactiveColor,
@@ -445,6 +450,7 @@ class ThemeData with Diagnosticable {
   static ThemeData lerp(ThemeData a, ThemeData b, double t) {
     return ThemeData.raw(
       brightness: t < 0.5 ? a.brightness : b.brightness,
+      extensions: t < 0.5 ? a.extensions : b.extensions,
       visualDensity: t < 0.5 ? a.visualDensity : b.visualDensity,
       resources: ResourceDictionary.lerp(a.resources, b.resources, t),
       accentColor: AccentColor.lerp(a.accentColor, b.accentColor, t),
@@ -509,8 +515,28 @@ class ThemeData with Diagnosticable {
     );
   }
 
+  /// Used to obtain a particular [ThemeExtension] from [extensions].
+  ///
+  /// Obtain with `FluentTheme.of(context).extension<MyThemeExtension>()`.
+  ///
+  /// See [extensions] for an interactive example.
+  T? extension<T>() => extensions[T] as T?;
+
+  /// Convert the [extensionsIterable] passed to [ThemeData.new] or [copyWith]
+  /// to the stored [extensions] map, where each entry's key consists of the extension's type.
+  static Map<Object, ThemeExtension<dynamic>> _themeExtensionIterableToMap(
+      Iterable<ThemeExtension<dynamic>> extensionsIterable) {
+    return Map<Object, ThemeExtension<dynamic>>.unmodifiable(<Object,
+        ThemeExtension<dynamic>>{
+      // Strangely, the cast is necessary for tests to run.
+      for (final ThemeExtension<dynamic> extension in extensionsIterable)
+        extension.type: extension as ThemeExtension<ThemeExtension<dynamic>>,
+    });
+  }
+
   ThemeData copyWith({
     Brightness? brightness,
+    Iterable<ThemeExtension<dynamic>>? extensions,
     VisualDensity? visualDensity,
     Typography? typography,
     AccentColor? accentColor,
@@ -558,6 +584,9 @@ class ThemeData with Diagnosticable {
       brightness: brightness ?? this.brightness,
       visualDensity: visualDensity ?? this.visualDensity,
       typography: this.typography.merge(typography),
+      extensions: extensions != null
+          ? _themeExtensionIterableToMap(extensions)
+          : this.extensions,
       accentColor: accentColor ?? this.accentColor,
       activeColor: activeColor ?? this.activeColor,
       inactiveColor: inactiveColor ?? this.inactiveColor,