+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **MINOR BREAK** `TextBox.clearGlobalKey` was remove, since it was not used
 - Add `AutoSuggestBox.unfocusedColor` and `TextFormBox.unfocusedColor`
 - Implement `displayInfoBar`, which shows an info bar as an overlay ([#673](https://github.com/bdlukaa/fluent_ui/issues/673))
+- Implement `ThemeData.extensions` ([#674](https://github.com/bdlukaa/fluent_ui/issues/674))
 
 ## 4.1.3